refactor (package): domain package 추가 - domain.car, domain.game, domain.strategy
