+++ src/main/java/racingcar/domain/car/CarName.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.car;
 
 import org.apache.commons.lang3.StringUtils;
