+++ src/main/java/racingcar/domain/car/CarPosition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.car;
 
 public class CarPosition {
