+++ src/main/java/racingcar/domain/strategy/RandomMovableStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.strategy;
 
 import java.util.Random;
