+++ src/test/java/racingcar/CarPositionTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.car.CarPosition;
 
 import static org.assertj.core.api.Assertions.*;
