+++ src/test/java/racingcar/RacingGameTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racingcar.domain.game.RacingGame;
+import racingcar.domain.game.RacingGameResult;
 
 import java.util.stream.Stream;
