+++ src/test/java/racingcar/RandomMovableStrategyTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.mockito.junit.jupiter.MockitoExtension;
+import racingcar.domain.strategy.RandomMovableStrategy;
 
 import java.util.Random;
